"Rationalized" dependencies chain, and now using "true" version numbers for every component.
I guess there *should* be some hard limit in how much fglrx can eschew catalyst, but I dare someone to prove that.
Also, this is not *really* using the 15.12 release.. but linux never got 16.xx, and at the end of the day this is still based on the same branch of the November driver.